refactor: namespace-ify logic, modernize enums, cleanup stdfunc
io.hpp:
- include <algorithm> and <string_view> for future use
- fix write() syscall to cast descriptor_t via uintptr_t

logic.hpp:
- wrap all logic symbols under `namespace logic`
- introduce anonymous namespace for constants and macros
- convert `item_t` and `actor_t` to scoped enum class with uint8_t backing
- rename functions: drop `logic$` prefix (e.g. logic$map$init → map$init)
- replace raw enum values with `item_t::` / `actor_t::` qualifiers
- ensure proper casting for char ↔ enum conversions
- cleanup player inventory, fight, map, and action handling accordingly

runtime.hpp:
- use `logic::` namespace in init(), iterate(), event()
- updated player lose check and map functions to new namespace style

stdfunc.hpp:
- remove unused headers <algorithm>, <ranges>, <span>, <type_traits>
- move back into `namespace stdfunc`
- mark utility functions as constexpr and FORCE_INLINE
- simplify convertNumberToString with static_cast and constexpr power()

system.hpp:
- add <array> include for upcoming use

+++ io.hpp
@@ -1,6 +1,8 @@
 #pragma once
 
+#include <algorithm>
 #include <cstdint>
+#include <string_view>
 
 #include "stdfunc.hpp"
 #include "system.hpp"
@@ -20,7 +22,8 @@ FORCE_INLINE void read( char* _buffer, uint8_t _bufferLength ) {
 }
 
 FORCE_INLINE void write( const char* _buffer, uint8_t _bufferLength ) {
-    syscall( system::call_t::write, ( uint8_t )descriptor_t::out,
+    syscall( system::call_t::write,
+             static_cast< uintptr_t >( descriptor_t::out ),
              std::bit_cast< uintptr_t >( _buffer ), _bufferLength );
 }
